refactor: move toTransferableNestings
